src/ai/llamaindex/document_loaders/requirements.txt
# Document Loader Dependencies for Haven Health Passport

# PDF Processing
pypdf>=3.17.0
pdfplumber>=0.10.0
pdf2image>=1.16.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10

# OCR
# Note: Tesseract must be installed separately
# Ubuntu/Debian: sudo apt-get install tesseract-ocr
# macOS: brew install tesseract

# Text Processing
chardet>=5.2.0
langdetect>=1.0.9

# Office Documents (future implementation)
# python-docx>=1.1.0
# openpyxl>=3.1.0

# Medical Imaging (future implementation)
# pydicom>=2.4.0

# HL7 Processing (future implementation)
# python-hl7>=0.4.5
# fhir.resources>=6.5.0

# Medical NLP (future implementation)
# scispacy>=0.5.0
# medspacy>=1.0.0

# Additional utilities
numpy>=1.24.0
python-magic>=0.4.27  # For file type detection
